+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Unreleased
+- /
+
+## [0.13.1] - 2024-09-13
 
 - Fix removing of potion brewing
 - Fix enchantments not being applied to the item in the enchantment table
@@ -104,6 +107,7 @@
 - Add `morejs.player.start_trading` event.
 
 <!-- Versions -->
+[0.13.1]: https://github.com/AlmostReliable/morejs/releases/tag/v1.21-neoforge-0.13.1
 [0.13.0]: https://github.com/AlmostReliable/morejs/releases/tag/v1.21-neoforge-0.13.0
 [0.12.0]: https://github.com/AlmostReliable/morejs/releases/tag/v1.21-neoforge-0.12.0
 [0.11.0]: https://github.com/AlmostReliable/morejs/releases/tag/v1.21-neoforge-0.11.0